stats and time
