+++ go.mod
@@ -1,3 +1,5 @@
-module smartSurvey
+module github.com/iurikman/smartSurvey
 
 go 1.22
+
+require github.com/gorilla/mux v1.8.1 // indirect
